move features into "features" folder and update include paths
