+++ src/core/MSSCore.vcxproj
@@ -105,32 +105,32 @@
   <ItemGroup>
     <ClCompile Include="controllers\BackendControllerBase\backendcontrollerbase.cpp" />
     <ClCompile Include="controllers\MountController\mountcontroller.cpp" />
-    <ClCompile Include="mounts\Mount\mount.cpp" />
-    <ClCompile Include="processing\FrameProcessor\frameprocessor.cpp" />
-    <ClCompile Include="processing\ProcessorBase\processorbase.cpp" />
-    <ClCompile Include="processing\YOLO\yolo.cpp" />
-    <ClCompile Include="sources\IClippableSource\ClipBufferBase\clipbufferbase.cpp" />
-    <ClCompile Include="sources\Source\source.cpp" />
-    <ClCompile Include="sources\Source\SourceMetadata\sourcemetadata.cpp" />
-    <ClCompile Include="sources\Source\SourcePreview\sourcepreview.cpp" />
-    <ClCompile Include="sources\Source\SourcePropertiesWidget\sourcepropertieswidget.cpp" />
+    <ClCompile Include="features\mounts\Mount\mount.cpp" />
+    <ClCompile Include="features\processors\FrameProcessor\frameprocessor.cpp" />
+    <ClCompile Include="features\processors\ProcessorBase\processorbase.cpp" />
+    <ClCompile Include="features\processors\YOLO\yolo.cpp" />
+    <ClCompile Include="features\sources\IClippableSource\ClipBufferBase\clipbufferbase.cpp" />
+    <ClCompile Include="features\sources\Source\source.cpp" />
+    <ClCompile Include="features\sources\Source\SourceMetadata\sourcemetadata.cpp" />
+    <ClCompile Include="features\sources\Source\SourcePreview\sourcepreview.cpp" />
+    <ClCompile Include="features\sources\Source\SourcePropertiesWidget\sourcepropertieswidget.cpp" />
     <QtMoc Include="controllers\BackendControllerBase\backendcontrollerbase.h" />
-    <QtMoc Include="mounts\Mount\mount.h" />
+    <QtMoc Include="features\mounts\Mount\mount.h" />
     <QtMoc Include="controllers\MountController\mountcontroller.h" />
     <ClInclude Include="interfaces\plugins\imountplugin.h" />
     <ClInclude Include="presets\Preset\preset.h" />
     <ClInclude Include="presets\SourcePreset\sourcepreset.h" />
-    <ClInclude Include="sources\IClippableSource\iclippablesource.h" />
-    <ClInclude Include="sources\IConfigurableSource\iconfigurablesource.h" />
-    <QtMoc Include="sources\Source\source.h" />
-    <ClInclude Include="sources\Source\SourceError\sourceerror.h" />
-    <QtMoc Include="sources\Source\SourcePropertiesWidget\sourcepropertieswidget.h" />
-    <QtMoc Include="sources\Source\SourcePreview\sourcepreview.h" />
-    <QtMoc Include="sources\Source\SourceMetadata\sourcemetadata.h" />
-    <QtMoc Include="sources\IClippableSource\ClipBufferBase\clipbufferbase.h" />
-    <QtMoc Include="processing\YOLO\yolo.h" />
-    <QtMoc Include="processing\ProcessorBase\processorbase.h" />
-    <QtMoc Include="processing\FrameProcessor\frameprocessor.h" />
+    <ClInclude Include="features\sources\IClippableSource\iclippablesource.h" />
+    <ClInclude Include="features\sources\IConfigurableSource\iconfigurablesource.h" />
+    <QtMoc Include="features\sources\Source\source.h" />
+    <ClInclude Include="features\sources\Source\SourceError\sourceerror.h" />
+    <QtMoc Include="features\sources\Source\SourcePropertiesWidget\sourcepropertieswidget.h" />
+    <QtMoc Include="features\sources\Source\SourcePreview\sourcepreview.h" />
+    <QtMoc Include="features\sources\Source\SourceMetadata\sourcemetadata.h" />
+    <QtMoc Include="features\sources\IClippableSource\ClipBufferBase\clipbufferbase.h" />
+    <QtMoc Include="features\processors\YOLO\yolo.h" />
+    <QtMoc Include="features\processors\ProcessorBase\processorbase.h" />
+    <QtMoc Include="features\processors\FrameProcessor\frameprocessor.h" />
     <QtMoc Include="controllers\ClipController\clipcontroller.h" />
     <QtMoc Include="controllers\MainController\maincontroller.h" />
     <QtMoc Include="controllers\PluginController\plugincontroller.h" />
