+++ src/core/controllers/MountController/mountcontroller.cpp
@@ -7,3 +7,8 @@ MountController::MountController(QObject *parent)
 MountController::~MountController()
 {}
 
+const Mount* MountController::byId(const QUuid& id) const
+{
+	// TODO: Implement the logic to find a Mount by its ID
+	return nullptr;
+}
