+++ src/core/features/sources/IClippableSource/ClipBufferBase/clipbufferbase.cpp
