+++ src/core/features/sources/IClippableSource/iclippablesource.h
