+++ src/core/features/sources/Source/source.cpp
@@ -2,7 +2,7 @@
 
 Source::Source(QByteArray hardwareId, QObject* parent)
 {
-	mId = hardwareId;
+	mId = QUuid::fromBytes(hardwareId); // Must be big endian to be reversible. TODO: make sure this works on all platforms
 }
 
 Source::Source(QObject *parent)
